Fixed: Showing already imported episodes as downloading in Series index
