+++ frontend/src/App/State/QueueAppState.ts
@@ -1,41 +1,9 @@
-import ModelBase from 'App/ModelBase';
-import Language from 'Language/Language';
-import { QualityModel } from 'Quality/Quality';
-import CustomFormat from 'typings/CustomFormat';
+import Queue from 'typings/Queue';
 import AppSectionState, {
   AppSectionFilterState,
   AppSectionItemState,
   Error,
 } from './AppSectionState';
-
-export interface StatusMessage {
-  title: string;
-  messages: string[];
-}
-
-export interface Queue extends ModelBase {
-  languages: Language[];
-  quality: QualityModel;
-  customFormats: CustomFormat[];
-  size: number;
-  title: string;
-  sizeleft: number;
-  timeleft: string;
-  estimatedCompletionTime: string;
-  status: string;
-  trackedDownloadStatus: string;
-  trackedDownloadState: string;
-  statusMessages: StatusMessage[];
-  errorMessage: string;
-  downloadId: string;
-  protocol: string;
-  downloadClient: string;
-  outputPath: string;
-  episodeHasFile: boolean;
-  seriesId?: number;
-  episodeId?: number;
-  seasonNumber?: number;
-}
 
 export interface QueueDetailsAppState extends AppSectionState<Queue> {
   params: unknown;
